Add a go http server for probers.
Adds an http server written in go that implements a stripped down restful
api to accept and display prober updates.

The intent is for the java probers to send json updates to the go server, which
will then display a running history of the status of probes (green for success
and red for failures).